Add shared code to a common module
